Fix Logics & update API to be async & code Refactor
