+++ event-buds-api/Service/EventInvitationService.py
@@ -9,32 +9,32 @@ class EventInvitationService:
     def __init__(self):
         self.repository = EventInvitationRepository.EventInvitationRepository()
 
-    def sendEventInvitation(self, invitation: EventInvitation):
+    async def sendEventInvitation(self, invitation: EventInvitation):
         try:
-            self.repository.sendEventInvitation(invitation)
+            await self.repository.sendEventInvitation(invitation)
             return "Success"
         except NameError as e:
             return e
 
-    def updateEventInvitation(self, invitation_id, message):
-        return self.repository.updateInvitation(invitation_id=invitation_id, message=message)
+    async def updateEventInvitation(self, invitation_id, message):
+        return await self.repository.updateInvitation(invitation_id=invitation_id, message=message)
 
-    def getUserInvitations(self, token):
+    async def getUserInvitations(self, token):
         payload = decodeJWT(token)
         user_id: int = payload.get("user_id")
-        return self.repository.getUserInvitation(user_id=user_id)
+        return await self.repository.getUserInvitation(user_id=user_id)
 
-    def getEventInvitations(self, event_id=0):
+    async def getEventInvitations(self, event_id=0):
         return self.repository.getEventInvitation(event_id)
 
-    def getEventHelpers(self, event_id=0):
-        return self.repository.getEventHelpers(event_id)
+    async def getEventHelpers(self, event_id=0):
+        return await self.repository.getEventHelpers(event_id)
 
-    def getEventGuests(self, event_id=0):
-        return self.repository.getEventGuests(event_id)
+    async def getEventGuests(self, event_id=0):
+        return await self.repository.getEventGuests(event_id)
 
-    def updateTaskAssignedInvitation(self, invitation_id):
-        return self.repository.updateTaskAssignedInvitation(invitation_id)
+    async def updateTaskAssignedInvitation(self, invitation_id):
+        return await self.repository.updateTaskAssignedInvitation(invitation_id)
 
-    def updateInvitationNotified(self, invitation_id):
-        return self.repository.updateInvitationNotified(invitation_id)
+    async def updateInvitationNotified(self, invitation_id):
+        return await self.repository.updateInvitationNotified(invitation_id)
