+++ event-buds-ui/src/pages/Public Event/PublicEvent.tsx
@@ -22,7 +22,7 @@ import { chevronDownCircleOutline } from "ionicons/icons";
 import { useContext, useEffect, useState } from "react";
 import { GetPublicEvents } from "../../api/eventApi";
 import Menu from "../../components/Menu";
-import PublicEventInfo from "../../components/PublicEventInfo";
+import PublicEventInfo from "../../components/EventInformation/PublicEventInfo";
 import { UserContext } from "../../context/UserContext";
 
 import "./publicEvent.css";
@@ -38,7 +38,9 @@ function PublicEvents() {
     setIsLoading(true);
     let result = await GetPublicEvents(token);
     if (result) {
-      setPublicEvents(result);
+      setPublicEvents(
+        result.filter((event: any) => new Date(event.REGENDDATE) > new Date())
+      );
     }
     setIsLoading(false);
   }
@@ -47,13 +49,13 @@ function PublicEvents() {
     const target = ev.target as HTMLIonSearchbarElement;
 
     query = target.value!.toLowerCase();
-
     setResults(
       publicEvents?.filter(
         (event: any) =>
-          event.EVENTTITLE?.toLowerCase().indexOf(query) > -1 ||
-          event?.DESCRIPTION?.toLowerCase().indexOf(query) > -1 ||
-          event.LOCATION?.toLowerCase().indexOf(query) > -1
+          new Date(event.REGENDDATE) > new Date() &&
+          (event.EVENTTITLE?.toLowerCase().indexOf(query) > -1 ||
+            event?.DESCRIPTION?.toLowerCase().indexOf(query) > -1 ||
+            event.LOCATION?.toLowerCase().indexOf(query) > -1)
       )
     );
   };
@@ -61,6 +63,7 @@ function PublicEvents() {
     if (userLoggedIn) {
       loadPublicEvents();
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [userLoggedIn]);
 
   useEffect(() => {
